Improve day 2 part 2 performance
Improve day 2 part 2 performance by limiting permutation of unsafe
reports checked.

If a report is safe, no permutations to check.

If a report is unsafe because of too large a diff, no permutations
to check. Because a safe report must have strictly increasing or
decreasing levels, if we ever see a diff too large, it is impossible
to make the report safe by removing only one level.

If a report is unsafe because of too small a diff (consecutive
identical levels) try with the second consecutive level removed. We
still have to check the rest of the report because there may be
other violations later on.

If a report is unsafe because of a direction change, try n + 1
permutations where n is the index where the direction change was
detected. These permutations are each with a single index removed,
beginning at index 0 and up to and including index n. We may be
able to limit this further, but we already see a nice performance
improvement from what we have here. So, we will check this in.