Make pathfinder hold reference to map instead of own
For part 2 implementation, we need the ability to create an
arbitrary number of sub-pathfinders to check if an obstacle
location would create a loop. To avoid having to clone our
map each time we do this, let's change the Pathfinder class
to hold a reference to the map, instead of owning it. Add
necessary lifetime annotations.