Make map reference immutable in pathfinder
Add optional obstacle location field to pathfinder. Subpathfinders
spawned by the primary pathfinder will use this when populating
their path, instead of relying on map modifications. This sets us
up for multithreading, as we won't have to worry about threads
reading the map as it's modified (the rust compiler wouldn't allow
this anyway)