Progress on day 9 part 2 implementation
I completed an implementation for day 9 part 2, but I misinterpreted
the spec, so the implementation is wrong. For my implementation, the
earliest block of blank spaces gets the latest possible file moved
into it, and each block of blank spaces is considered exactly once
for swapping with a file. The spec is the other way around. Each
file should be considered at most once to swap into some blank
space. But a given blank space can be considered each time a file
is up to be swapped. Saving for posterity.