Nice runtime performance improvement for day 9 part 2
To improve performance of the day 9 part 2 implementation:
- Use a vector to store file block lengths. The index into the
vector represents the file id. Since we remove only from the
end, we know the file id -> index mapping is preserved.
- Use a vector to store blank block lengths. Here, we don't
care about mapping to an id because blank blocks do not need
an id.

When we perform block compression, we start with the latest
file added. We then linear search across our blank block
vector until we find a blank block that is large enough to
hold the file and is before the file in memory (if no such
blank block is found, we simply continue to the next latest
file, and the first file does not ever move). We swap the
blank block region and the file region in our disk memory,
then we update our blank block region in our blank vector
to reflect its reduced size. If its size is now 0, we remove
it from the blank vector.

Note we do not care about the blank region that got swapped
into the file's previous location. Because we only check each
file once and we check them in descending order, we know we
will never use the blank region created when a file is swapped.
So, we do not bother tracking it.