Greatly improve day 11 implementation performance
This commit includes a working solution for day 11 part 2. To
improve performance, we store a map of stone counts instead of
individual stones. On a blink, when we transform a stone, we
update the count map using its prior count for its one or two
children created by the transformation. This allows us to avoid
duplicate calculations when we have duplicate stones at a given
step.