Performance improvement and part 2 setup
Setup infrastructure to run part 2 solution. Part 1 and part 2
share a solution, but part 2 requires a much quicker implementation
to complete in a reasonable time.

Improve performance by checking exactly one case for each amount of
cheap button presses. We can check if, after the num of cheap
button presses, we are now the same multiple away from the prize in
both the x and y direction, with respect to the x and y steps in
the pricier button.