Input processing and horizontal shifts for day 15 part 2
Implement input processing to double the width of everything in
the warehouse other than the robot for day 15 part 2. Add logic to
shift double wide boxes horizontally as the robot moves. Logic to
move boxes vertically will be in a later commit.

Add manual test case for playing around with robot movement in
double wide warehouse.