Add grid util to find a singular occuring char in map
Make find_robot method from Lanternfish a grid util method that
takes in a char and a reference to a 2d char vector, and returns
the location of that char. It is expected the char appears only
once in the map. If it does not appear, the program panics. If
it appears more than once, do not use this method :)