Initial day 17 part 2 working implementation
Implement method to take in a program and compute the minimal
register A value that would lead to a computer outputting a copy
of the program when it runs. This implementation starts by
finding the smallest number that will output the final number of
the program, then multiplies that by 8 and begins looking there
for the smallest number that will output the final two numbers of
the program, etc etc.

To be honest I don't understand why this works. I thought that
each subsequent number we find must be in the range
[prior * 8, prior * 8 + 8)

But this did not hold true for each octal digit of the solution
value.